Rename validity assertions to contain 'structure' in the name
