+++ packages/jest-structure/index.js
@@ -1,6 +1,6 @@
 module.exports = {
-  toBeValid: require('./src/assertions/toBeValid'),
-  toBeInvalid: require('./src/assertions/toBeInvalid'),
+  toBeValidStructure: require('./src/assertions/toBeValidStructure'),
+  toBeInvalidStructure: require('./src/assertions/toBeInvalidStructure'),
   toHaveInvalidAttribute: require('./src/assertions/toHaveInvalidAttribute'),
   toHaveInvalidAttributes: require('./src/assertions/toHaveInvalidAttributes'),
 };
